python_api_service/requirements-simple.txt
# Minimal requirements - Chỉ cài những gì cần thiết

# Web Framework
fastapi
uvicorn[standard]
python-multipart

# PDF Processing (sẽ tự động cài phiên bản mới nhất tương thích)
PyMuPDF>=1.24.0
pdf2image
Pillow

# OCR
pytesseract

# Utilities
python-dotenv

# Lưu ý: Transformers và Torch KHÔNG cần thiết cho chức năng cơ bản
# Hệ thống vẫn hoạt động tốt mà không cần AI NLP phức tạp
